+++ configurations/flyway/src/main/java/io/micronaut/configuration/dbmigration/flyway/FlywayStartupEventListener.java
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-package io.micronaut.dbmigration.flyway;
+package io.micronaut.configuration.dbmigration.flyway;
 
 import io.micronaut.context.ApplicationContext;
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.context.event.StartupEvent;
 import io.micronaut.runtime.event.annotation.EventListener;
-import io.micronaut.scheduling.annotation.Async;
 import org.flywaydb.core.Flyway;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -29,37 +28,36 @@ import javax.inject.Singleton;
 import java.util.Collection;
 
 /**
- * Asynchronous listener for {@link StartupEvent} to run flyway operations.
+ * Synchronous listener for {@link StartupEvent} to run flyway operations.
  *
  * @author Iván López
  * @since 1.1
  */
 @Requires(beans = Flyway.class)
 @Singleton
-class FlywayAsyncStartupEventListener extends AbstractFlyway {
+class FlywayStartupEventListener extends AbstractFlyway {
 
-    private static final Logger LOG = LoggerFactory.getLogger(FlywayAsyncStartupEventListener.class);
+    private static final Logger LOG = LoggerFactory.getLogger(FlywayStartupEventListener.class);
 
     /**
      * @param applicationContext            The application context
      * @param flywayConfigurationProperties Collection of Flyway configuration properties
      */
-    public FlywayAsyncStartupEventListener(ApplicationContext applicationContext,
-                                           Collection<FlywayConfigurationProperties> flywayConfigurationProperties) {
+    public FlywayStartupEventListener(ApplicationContext applicationContext,
+                                      Collection<FlywayConfigurationProperties> flywayConfigurationProperties) {
         super(applicationContext, flywayConfigurationProperties);
     }
 
     /**
-     * Runs Flyway migrations asynchronously.
+     * Runs Flyway migrations synchronously.
      *
      * @param event Server startup event
      */
-    @Async
     @EventListener
     public void onStartup(StartupEvent event) {
         if (LOG.isTraceEnabled()) {
             LOG.trace("Executing flyway event listener");
         }
-        run(true);
+        run(false);
     }
 }
