+++ configurations/flyway/src/test/groovy/io/micronaut/configuration/dbmigration/flyway/management/endpoint/FlywayEndpointSpec.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.micronaut.dbmigration.flyway.management.endpoint
+package io.micronaut.configuration.dbmigration.flyway.management.endpoint
 
 import io.micronaut.context.ApplicationContext
 import io.micronaut.context.env.Environment
